refactor: MockLLMClient 移出 utils，消除 utils→agents 反向依赖
- MockLLMClient（~290 行产品级 mock 应答）迁至 services/mock_llm.py，
  utils/llm.py 回归纯 LLM 客户端抽象（489→199 行），
  get_llm_client 惰性导入 mock，无循环依赖
- output_style.py 从 agents/ 迁至 core/（只依赖 core.schemas，
  属横切关注点），utils/llm.py 不再 import agents 任何符号
- 同步更新 17 个文件的导入路径（2 个 src + 15 个测试）
- 顺带修复 base.py/cache.py 的 UP047（PEP 695 类型参数）

验证: ruff 0 错误; pytest 607 passed / 3 skipped
