+++ src/stockresearch/api/routes/advisor.py
@@ -7,9 +7,9 @@
 from fastapi import APIRouter, Body, Depends
 from sqlalchemy.orm import Session
 
-from stockresearch.agents.output_style import output_style_scope
 from stockresearch.api.deps import get_current_user
 from stockresearch.api.llm_deps import llm_from_headers
+from stockresearch.core.output_style import output_style_scope
 from stockresearch.core.schemas import AssetAllocationOut, AssetAllocationRequest
 from stockresearch.db.models import User
 from stockresearch.db.session import get_db
