+++ src/stockresearch/core/output_style.py
