+++ src/stockresearch/services/cache.py
@@ -16,9 +16,6 @@ import json
 import time
 from collections import OrderedDict
 from collections.abc import Callable
-from typing import TypeVar
-
-T = TypeVar("T")
 
 _MAX_MEMORY_ENTRIES = 1000
 _MAX_FACTORY_ENTRIES = 200
@@ -84,7 +81,7 @@ class CacheService:
         _factory_store.clear()
 
 
-def get_cached(key: str, ttl_sec: float, factory: Callable[[], T]) -> T:
+def get_cached[T](key: str, ttl_sec: float, factory: Callable[[], T]) -> T:
     now = time.monotonic()
     entry = _factory_store.get(key)
     if entry is not None and now - entry[0] < ttl_sec:
@@ -96,7 +93,7 @@ def get_cached(key: str, ttl_sec: float, factory: Callable[[], T]) -> T:
     return value
 
 
-def peek_cached(key: str, ttl_sec: float) -> T | None:
+def peek_cached(key: str, ttl_sec: float) -> object | None:
     """Return a cached factory value without invoking the factory."""
     now = time.monotonic()
     entry = _factory_store.get(key)
