+++ tests/agents/test_battle_vote.py
@@ -4,7 +4,7 @@ import pytest
 
 from stockresearch.agents.research.debate import iter_battle_vote_events
 from stockresearch.core.schemas import DimensionResult
-from stockresearch.utils.llm import MockLLMClient
+from stockresearch.services.mock_llm import MockLLMClient
 
 
 @pytest.mark.asyncio
