+++ tests/agents/test_chat_sync_stream_parity.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 from stockresearch.agents.orchestrator.graph import Orchestrator
 from stockresearch.agents.orchestrator.stream import run_chat_stream
-from stockresearch.agents.output_style import output_style_scope
+from stockresearch.core.output_style import output_style_scope
 from stockresearch.db.models import User
