+++ tests/agents/test_multi_round_debate.py
@@ -3,7 +3,7 @@
 import pytest
 
 from stockresearch.agents.research.debate import run_multi_round_debate
-from stockresearch.utils.llm import MockLLMClient
+from stockresearch.services.mock_llm import MockLLMClient
 
 
 @pytest.mark.asyncio
