+++ tests/agents/test_secondary_block_injection.py
@@ -10,7 +10,7 @@ from stockresearch.agents.orchestrator.react_agent import OrchestratorAgent
 from stockresearch.core.schemas import ModeSettingsOut
 from stockresearch.db.models import User
 from stockresearch.services.chat.scope import build_chat_context_scope
-from stockresearch.utils.llm import MockLLMClient
+from stockresearch.services.mock_llm import MockLLMClient
 
 _MESSAGE = "介绍一下半导体和新能源的区别"
