+++ tests/test_llm_settings.py
@@ -7,8 +7,8 @@ from stockresearch.api.app import create_app
 from stockresearch.api.llm_deps import merge_llm_settings
 from stockresearch.core.llm_config import LlmOverrides, resolve_chat_completions_url
 from stockresearch.core.schemas import LlmUserSettings
+from stockresearch.services.mock_llm import MockLLMClient
 from stockresearch.utils.llm import (
-    MockLLMClient,
     OpenAICompatibleClient,
     _httpx_client_kwargs,
     get_llm_client,
